+++ src/hoc/withErrorHandler/withErrorHandler.js
@@ -1,29 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
+
+import useErrorHandler from '../../hooks/useErrorHandler';
 
 import Modal from '../../components/UI/Modal/Modal';
 
 const withErrorHandler = (WrappedComponent, axios) => {
   return props => {
-    const [error, setError] = useState(null);
-
-    const requestInterceptor = axios.interceptors.request.use(req => {
-      setError(null);
-      return req;
-    });
-    const responseInterceptor = axios.interceptors.response.use(res => res, err => {
-      setError(err);
-    });
-
-    useEffect(() => {
-      return () => {
-        axios.interceptors.request.eject(requestInterceptor);
-        axios.interceptors.response.eject(responseInterceptor);
-      }
-    }, [requestInterceptor, responseInterceptor]);
-
-    const errorConfirmedHandler = () => {
-      setError(null);
-    }
+    const [error, errorConfirmedHandler] = useErrorHandler(axios);
 
     return (
       <React.Fragment>
